update authorizer structure
