+++ src/views/Authorizer/index.tsx
@@ -3,6 +3,7 @@ import './index.css'
 import { Suspense } from 'react'
 import { Navigate, Route, Routes } from 'react-router'
 
+import { PolicyEvaluatorContextProvider } from '../../services/PolicyEvaluatorContextProvider'
 import Frame from './Authorizer'
 import Authorizer from './Authorizer'
 import AuthorizerApiDocs from './pages/ApiDocs'
@@ -12,18 +13,20 @@ import Modules from './pages/Modules'
 const PolicyInstanceDetails = () => {
   return (
     <Suspense fallback={<Frame />}>
-      <Routes>
-        <Route element={<Authorizer />} path="evaluator">
-          <Route element={<Evaluator />} path="" />
-        </Route>
-        <Route element={<Authorizer />} path="modules">
-          <Route element={<Modules />} path="" />
-        </Route>
-        <Route element={<Authorizer />} path="docs">
-          <Route element={<AuthorizerApiDocs />} path="" />
-        </Route>
-        <Route element={<Navigate replace to="modules" />} index />
-      </Routes>
+      <PolicyEvaluatorContextProvider>
+        <Routes>
+          <Route element={<Authorizer />} path="evaluator">
+            <Route element={<Evaluator />} path="" />
+          </Route>
+          <Route element={<Authorizer />} path="modules">
+            <Route element={<Modules />} path="" />
+          </Route>
+          <Route element={<Authorizer />} path="docs">
+            <Route element={<AuthorizerApiDocs />} path="" />
+          </Route>
+          <Route element={<Navigate replace to="modules" />} index />
+        </Routes>
+      </PolicyEvaluatorContextProvider>
     </Suspense>
   )
 }
