+++ src/views/Authorizer/pages/Evaluator/EvaluatorConent/DecisionTreeTable/styles.tsx
