+++ src/views/Authorizer/pages/Evaluator/EvaluatorConent/PolicyCommonFields/Decisions.tsx
@@ -1,9 +1,9 @@
-import Input from '../../../../../../../components/common/Input'
-import { Row } from '../../../../../../../components/common/Row'
+import Input from '../../../../../../components/common/Input'
+import { Row } from '../../../../../../components/common/Row'
 import {
   useCommonPolicyEvaluatorContext,
   usePolicyEvaluatorErrorContext,
-} from '../../../../../../../services/PolicyEvaluatorContextProvider/hooks'
+} from '../../../../../../services/PolicyEvaluatorContextProvider/hooks'
 import { FieldContainer } from '../styles'
 import { SubField } from '../Subfield'
