+++ src/views/Authorizer/pages/Evaluator/EvaluatorConent/PolicyIsFields/index.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useMemo } from 'react'
 
-import Label from '../../../../../../../components/common/Label'
-import { Row } from '../../../../../../../components/common/Row'
-import Select from '../../../../../../../components/common/Select'
-import { useIsPolicyEvaluatorContext } from '../../../../../../../services/PolicyEvaluatorContextProvider/hooks'
-import { ApiModule } from '../../../../../../../types/authorizer'
+import Label from '../../../../../../components/common/Label'
+import { Row } from '../../../../../../components/common/Row'
+import Select from '../../../../../../components/common/Select'
+import { useIsPolicyEvaluatorContext } from '../../../../../../services/PolicyEvaluatorContextProvider/hooks'
+import { ApiModule } from '../../../../../../types/authorizer'
 import { Decisions } from '../PolicyCommonFields/Decisions'
 import { ResourceContext } from '../PolicyCommonFields/ResourceContext'
 import { FieldContainer, SectionTitleContainer } from '../styles'
